+++ examples/angular/src/app/app.component.ts
@@ -1,30 +1,9 @@
-import { Component } from '@angular/core'
-import { BalOptionValue } from '@baloise/ui-library-next'
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
+  templateUrl: './app.component.html'
 })
 export class AppComponent {
-  isChecked = true
-  loading = false
-  options: BalOptionValue<number>[] = []
-
-  onChange(event: CustomEvent<boolean>) {
-    console.log('on change', event.detail)
-  }
-
-  onSelectInput(event: CustomEvent<string>) {
-    this.loading = true
-    fetch('https://swapi.dev/api/people/?search=' + event.detail)
-      .then(res => res.json())
-      .then(data => {
-        this.options = data.results.map((person: any, index: number) => ({
-          value: index,
-          text: person.name,
-        }))
-        this.loading = false
-      })
-  }
+  title = 'angular';
 }