+++ packages/library/src/components.d.ts
@@ -228,6 +228,10 @@ export namespace Components {
     }
     interface BalCheckbox {
         /**
+          * The tabindex of the control.
+         */
+        "balTabindex": number;
+        /**
           * If `true`, the checkbox is selected.
          */
         "checked": boolean;
@@ -281,6 +285,10 @@ export namespace Components {
     interface BalDataValue {
     }
     interface BalDatepicker {
+        /**
+          * The tabindex of the control.
+         */
+        "balTabindex": number;
         /**
           * Closes the datepicker dropdown after selection
          */
@@ -338,7 +346,7 @@ export namespace Components {
          */
         "triggerIcon": boolean;
         /**
-          * Selected option value.
+          * Selected date. Could also be passed as a string, which gets transformed to js date object.
          */
         "value": Date;
     }
@@ -351,6 +359,10 @@ export namespace Components {
           * If `true` the field spans over the whole width.
          */
         "expanded": boolean;
+        /**
+          * If `true` the dropdown content has a fixed width
+         */
+        "fixedContentWidth": boolean;
         /**
           * Returns the `HTMLDivElement` of the content element
          */
@@ -516,6 +528,10 @@ export namespace Components {
          */
         "autoComplete": boolean;
         /**
+          * The tabindex of the control.
+         */
+        "balTabindex": number;
+        /**
           * If `true` the input gets a clickable cursor style
          */
         "clickable": boolean;
@@ -658,6 +674,10 @@ export namespace Components {
     }
     interface BalRadio {
         /**
+          * The tabindex of the control.
+         */
+        "balTabindex": number;
+        /**
           * If `true`, the radio is selected.
          */
         "checked": boolean;
@@ -706,6 +726,10 @@ export namespace Components {
     }
     interface BalSelect {
         /**
+          * The tabindex of the control.
+         */
+        "balTabindex": number;
+        /**
           * Sets the value to null and resets the value of the input.
          */
         "clear": () => Promise<void>;
@@ -722,6 +746,10 @@ export namespace Components {
          */
         "expanded": boolean;
         /**
+          * Defines the placeholder of the input filter element.
+         */
+        "filterPlaceholder": string;
+        /**
           * Set this to `true` when the component is placed on a dark background.
          */
         "inverted": boolean;
@@ -730,6 +758,14 @@ export namespace Components {
          */
         "loading": boolean;
         /**
+          * If `true` multiple option can be selected
+         */
+        "multiple": boolean;
+        /**
+          * If `true` the filtering of the options is done outside of the component.
+         */
+        "noFilter": boolean;
+        /**
           * Opens the dropdown
          */
         "open": () => Promise<void>;
@@ -737,10 +773,6 @@ export namespace Components {
           * Defines the placeholder of the input element.
          */
         "placeholder": string;
-        /**
-          * If `true` the filtering of the options is done outside of the component.
-         */
-        "remote": boolean;
         /**
           * Defines the height of the dropdown list.
          */
@@ -762,11 +794,15 @@ export namespace Components {
          */
         "typeahead": boolean;
         /**
-          * Selected option value.
+          * Selected option values. Could also be passed as a string, which gets transformed.
          */
-        "value": string;
+        "value": string[];
     }
     interface BalSelectOption {
+        /**
+          * If `true` the option has a checkbox
+         */
+        "checkbox": boolean;
         /**
           * If `true` the option is focused
          */
@@ -1462,6 +1498,10 @@ declare namespace LocalJSX {
     }
     interface BalCheckbox {
         /**
+          * The tabindex of the control.
+         */
+        "balTabindex"?: number;
+        /**
           * If `true`, the checkbox is selected.
          */
         "checked"?: boolean;
@@ -1523,6 +1563,10 @@ declare namespace LocalJSX {
     interface BalDataValue {
     }
     interface BalDatepicker {
+        /**
+          * The tabindex of the control.
+         */
+        "balTabindex"?: number;
         /**
           * Closes the datepicker dropdown after selection
          */
@@ -1592,7 +1636,7 @@ declare namespace LocalJSX {
          */
         "triggerIcon"?: boolean;
         /**
-          * Selected option value.
+          * Selected date. Could also be passed as a string, which gets transformed to js date object.
          */
         "value"?: Date;
     }
@@ -1601,6 +1645,10 @@ declare namespace LocalJSX {
           * If `true` the field spans over the whole width.
          */
         "expanded"?: boolean;
+        /**
+          * If `true` the dropdown content has a fixed width
+         */
+        "fixedContentWidth"?: boolean;
         /**
           * If `true` the dropdown content is open.
          */
@@ -1754,6 +1802,10 @@ declare namespace LocalJSX {
          */
         "autoComplete"?: boolean;
         /**
+          * The tabindex of the control.
+         */
+        "balTabindex"?: number;
+        /**
           * If `true` the input gets a clickable cursor style
          */
         "clickable"?: boolean;
@@ -1906,6 +1958,10 @@ declare namespace LocalJSX {
     }
     interface BalRadio {
         /**
+          * The tabindex of the control.
+         */
+        "balTabindex"?: number;
+        /**
           * If `true`, the radio is selected.
          */
         "checked"?: boolean;
@@ -1962,6 +2018,10 @@ declare namespace LocalJSX {
     }
     interface BalSelect {
         /**
+          * The tabindex of the control.
+         */
+        "balTabindex"?: number;
+        /**
           * If `true` the component is diabled.
          */
         "disabled"?: boolean;
@@ -1970,6 +2030,10 @@ declare namespace LocalJSX {
          */
         "expanded"?: boolean;
         /**
+          * Defines the placeholder of the input filter element.
+         */
+        "filterPlaceholder"?: string;
+        /**
           * Set this to `true` when the component is placed on a dark background.
          */
         "inverted"?: boolean;
@@ -1977,6 +2041,14 @@ declare namespace LocalJSX {
           * If `true` the component shows a loading spinner and sets the input to readonly.
          */
         "loading"?: boolean;
+        /**
+          * If `true` multiple option can be selected
+         */
+        "multiple"?: boolean;
+        /**
+          * If `true` the filtering of the options is done outside of the component.
+         */
+        "noFilter"?: boolean;
         /**
           * Emitted when the input loses focus.
          */
@@ -1988,7 +2060,7 @@ declare namespace LocalJSX {
         /**
           * Emitted when a option got selected.
          */
-        "onBalChange"?: (event: CustomEvent<string>) => void;
+        "onBalChange"?: (event: CustomEvent<string[]>) => void;
         /**
           * Emitted when the input got clicked.
          */
@@ -2010,10 +2082,6 @@ declare namespace LocalJSX {
          */
         "placeholder"?: string;
         /**
-          * If `true` the filtering of the options is done outside of the component.
-         */
-        "remote"?: boolean;
-        /**
           * Defines the height of the dropdown list.
          */
         "scrollable"?: number;
@@ -2022,11 +2090,15 @@ declare namespace LocalJSX {
          */
         "typeahead"?: boolean;
         /**
-          * Selected option value.
+          * Selected option values. Could also be passed as a string, which gets transformed.
          */
-        "value"?: string;
+        "value"?: string[];
     }
     interface BalSelectOption {
+        /**
+          * If `true` the option has a checkbox
+         */
+        "checkbox"?: boolean;
         /**
           * If `true` the option is focused
          */