+++ packages/library/src/components/bal-dropdown/bal-dropdown.tsx
@@ -1,4 +1,4 @@
-import { Component, h, Host, Listen, Method, Prop, Element, Event, EventEmitter } from '@stencil/core'
+import { Component, h, Host, Listen, Method, Prop, Element, Event, EventEmitter, State } from '@stencil/core'
 
 @Component({
   tag: 'bal-dropdown',
@@ -14,6 +14,8 @@ export class Dropdown {
 
   @Element() element!: HTMLElement
 
+  @State() isDropDownContentUp = false
+
   /**
    * Limit the height of the dropdown content. Pass the amount of pixel.
    */
@@ -23,6 +25,11 @@ export class Dropdown {
    * If `true` the field spans over the whole width.
    */
   @Prop() expanded: boolean = false
+
+  /**
+   * If `true` the dropdown content has a fixed width
+   */
+  @Prop() fixedContentWidth: boolean = false
 
   /**
    * If `true` the dropdown content is open.
@@ -109,7 +116,18 @@ export class Dropdown {
   async clickOnOutside(event: UIEvent) {
     if (!this.element.contains(event.target as any) && this.isActive) {
       await this.toggle()
+      this.calcIsDropDownContentUp()
     }
+  }
+
+  @Listen('resize', { target: 'window' })
+  async resizeHandler() {
+    this.calcIsDropDownContentUp()
+  }
+
+  @Listen('scroll', { target: 'window' })
+  async scrollHandler() {
+    this.calcIsDropDownContentUp()
   }
 
   get contentStyle() {
@@ -124,9 +142,14 @@ export class Dropdown {
     }
   }
 
-  get isUp(): boolean {
+  private calcIsDropDownContentUp() {
     const box = this.element.getBoundingClientRect()
-    return window.innerHeight - box.top < window.innerHeight / 2
+    const clientHeight = this.contentElement?.clientHeight || 250
+    this.isDropDownContentUp = window.innerHeight < box.bottom + clientHeight + 50
+  }
+
+  componentWillRender() {
+    this.calcIsDropDownContentUp()
   }
 
   render() {
@@ -136,8 +159,9 @@ export class Dropdown {
           class={{
             'dropdown': true,
             'is-active': this.isActive,
+            'has-fixed-content-width': this.fixedContentWidth,
             'is-expanded': this.expanded,
-            'is-up': this.isUp,
+            'is-up': this.isDropDownContentUp,
           }}>
           <div class="dropdown-trigger">
             <slot name="trigger" />